+++ aaem/components/wind_power.py
@@ -20,9 +20,9 @@ yaml = {'enabled': False,
         'lifetime': 'ABSOLUTE DEFAULT',
         'start year': 'ABSOLUTE DEFAULT',
         'average load limit': 100.0,
-        'percent generation to offset': .30,
+        'percent generation to offset': 1.00,
         'data':'IMPORT',
-        'minimum wind class': 3,
+         #'minimum wind class': 3,
         'wind cost': 'UNKNOWN',
         'secondary load': True,
         'secondary load cost': 200000,
@@ -140,21 +140,28 @@ def wind_preprocess (ppo):
                                        'Load','Certainty',
                                        'Estimated Generation','Estimated Cost',
                                        'Note','Resource Note'])
+                                       
+    try:
+        intertie = read_csv(os.path.join(ppo.data_dir,
+                            "wind_data_interties.csv"),
+                            comment = '#',
+                            index_col = 0).ix[ppo.com_id+"_intertie"]
+        intertie = int(intertie['Highest Wind Class on Intertie'])
+    except KeyError:
+        intertie = 0
+    
+    try:
+        if intertie > int(potential['Assumed Wind Class']):
+            potential.ix['Assumed Wind Class'] = intertie
+            ppo.diagnostics.add_note("wind", 
+                    "Wind class updated to max on intertie")
+        
+    except KeyError:
+        pass
+    
     assumptions = read_csv(os.path.join(ppo.data_dir,
                                 "wind_class_assumptions.csv"),
                            comment = '#',index_col = 0)
-    
-    try:
-        diesel = read_csv(os.path.join(ppo.data_dir,
-                                "diesel_data.csv"),
-                           comment = '#',index_col = 0).ix[ppo.com_id]
-    #~ print '\n\n\n\n'
-    #~ print diesel
-    #~ print '\n\n\n\n'
-    
-        hr = str(diesel['Waste Heat Recovery Opperational'])
-    except KeyError:
-        hr = 'nan'
     
     try:
         capa = assumptions.ix[int(float(potential.ix['Assumed Wind Class']))]
@@ -171,7 +178,6 @@ def wind_preprocess (ppo):
     fd.write("key,value\n")
     fd.write("existing wind," + str(existing) +'\n')
     fd.write('assumed capacity factor,' +str(capa) +'\n')
-    fd.write('Heat Recovery Opperational,' + str(hr) +'\n')
     fd.close()
 
     #~ df = concat([ww_d,ww_a])
@@ -199,7 +205,8 @@ raw_data_files = ['wind_class_assumptions.csv',
                   'wind_costs.csv',
                   "wind_data_existing.csv",
                   "wind_data_potential.csv",
-                  "diesel_data.csv"]
+                  "diesel_data.csv",
+                  'wind_data_interties.csv']
 
 ## list of wind preprocessing functions
 preprocess_funcs = [wind_preprocess, copy_wind_cost_table]
@@ -222,10 +229,11 @@ def component_summary (coms, res_dir):
             wind = coms[c]['model'].comps_used['wind power']
             
             average_load = wind.average_load
-            potential = wind.comp_specs['data']['Wind Potential']
+            existing_load = wind.comp_specs['data']['existing wind']
             wind_class = float(wind.comp_specs['data']['Assumed Wind Class']) 
             proposed_load =  wind.load_offset_proposed
-            
+            cap_fac = float(wind.comp_specs['data']['assumed capacity factor'])
+            heat_rec_opp = wind.comp_specs['data']['Heat Recovery Opperational']
             try:
                 #~ offset = wind.load_offset_proposed
                 net_gen_wind = wind.net_generation_wind
@@ -236,8 +244,7 @@ def component_summary (coms, res_dir):
                 
                 
                 diesel_red = wind.reduction_diesel_used
-                cap_fac = float(wind.comp_specs['data']\
-                                    ['assumed capacity factor'])
+                
                 eff = wind.cd["diesel generation efficiency"]
                 
             except AttributeError:
@@ -245,19 +252,33 @@ def component_summary (coms, res_dir):
                 net_gen_wind = 0
                 decbb = 0
                 
-                cap_fac = 0
                 loss_heat = 0
                 
                 diesel_red = 0
                 eff = wind.cd["diesel generation efficiency"]    
+                
+            try:
+                red_per_year = net_gen_wind / eff
+            except ZeroDivisionError:
+                red_per_year = 0
             
-            l = [c, average_load, potential, wind_class, proposed_load,
-                 net_gen_wind, decbb, loss_heat, diesel_red, 
-                 cap_fac, eff,
-                 wind.get_NPV_benefits(),
-                 wind.get_NPV_costs(),
-                 wind.get_NPV_net_benefit(),
-                 wind.get_BC_ratio()
+            l = [c, 
+                wind_class, 
+                average_load, 
+                proposed_load,
+                existing_load,
+                cap_fac,
+                net_gen_wind,
+                decbb, 
+                loss_heat, 
+                heat_rec_opp,
+                diesel_red, 
+                red_per_year,
+                eff,
+                wind.get_NPV_benefits(),
+                wind.get_NPV_costs(),
+                wind.get_NPV_net_benefit(),
+                wind.get_BC_ratio()
             ]
             out.append(l)
         except (KeyError,AttributeError) as e:
@@ -265,22 +286,24 @@ def component_summary (coms, res_dir):
             pass
         
     data = DataFrame(out,columns = \
-       ['community',
-        'average load [kw]',
-        'resource potential',
-        'assumed wind class',
-        'load offset proposed [kW]',
-        'Net Generation [kWh]',
-        'Diesel Equivalent Captured by Boilers [gal]',
-        'Loss of Recovered Heat[gal]',
-        'Net reduction reduction Diesel[gal]',
+       ['Community',
+        'Assumed Wind Class',
+        'Average Load [kw]',
+        'Wind Capacity Proposed [kW]',
+        'Existing Wind Capacity [kW]',
         'Assumed Capacity Factor [%]',
-        'diesel generator efficiency',
+        'Net Generation [kWh]',
+        'Heating Oil Equivalent Captured by Seconday Load [gal]',
+        'Loss of Recovered Heat[gal]',
+        'Heat Recovery Opperational',
+        'Net in Heating Oil Consumption [gal]',
+        'Reduction in Utility Diesel Consumed per year',
+        'Diesel Denerator Efficiency',
         'NPV benefits [$]',
         'NPV Costs [$]',
         'NPV Net benefit [$]',
         'Benefit Cost Ratio']
-                    ).set_index('community')#.round(2)
+                    ).set_index('Community')#.round(2)
     f_name = os.path.join(res_dir,
                 'wind_power_summary.csv')
     fd = open(f_name,'w')
@@ -350,10 +373,10 @@ class WindPower(AnnualSavings):
         
         #~ #~ print self.comp_specs['data']['Assumed Wind Class'] 
         # ??? some kind of failure message
-        if self.average_load > self.comp_specs['average load limit'] and\
-            float(self.comp_specs['data']['Assumed Wind Class']) > \
-                self.comp_specs['minimum wind class'] and \
-                self.load_offset_proposed > 0:
+        if self.average_load > self.comp_specs['average load limit']: #and\
+            #~ float(self.comp_specs['data']['Assumed Wind Class']) > \
+                #~ self.comp_specs['minimum wind class'] and \
+                #~ self.load_offset_proposed > 0:
         # if the average load is greater that the lower limit run this component
         # else skip    
             
@@ -436,8 +459,7 @@ class WindPower(AnnualSavings):
         offset = self.average_load*\
                 self.comp_specs['percent generation to offset']
         #~ self.comp_specs['data']['existing wind'] = 0
-        if self.comp_specs['data']['Wind Potential'] in ['H','M'] and \
-           int(float(self.comp_specs['data']['existing wind'])) < \
+        if int(float(self.comp_specs['data']['existing wind'])) < \
                 (round(offset/25) * 25): # ???
             #~ print "True"
             self.load_offset_proposed = round(offset/25) * 25 - \
@@ -526,9 +548,9 @@ class WindPower(AnnualSavings):
         """ 
              calulate the somthing ???
         """
-        hr_used = self.comp_specs['data']['Heat Recovery Opperational']
+        hr_used = self.cd['heat recovery operational']
         self.loss_heat_recovery = 0
-        if hr_used == 'Yes': 
+        if hr_used:# == 'Yes': 
             self.loss_heat_recovery = self.electric_diesel_reduction * \
             self.comp_specs['percent heat recovered']
         #~ print 'self.loss_heat_recovery',self.loss_heat_recovery